add -label flag to <account add> <account set> > add ... -label XXX: to add an account labeled XXX > set ... -label XXX: to set an account's label to XXX > list ...: will show account's label too
