+++ cmd/params.go
@@ -52,12 +52,14 @@ var (
 	listFlags = []cli.Flag{
 		utils.AccountVerboseFlag,
 		utils.AccountFileFlag,
+		utils.AccountLabelFlag,
 	}
 
 	setFlags = []cli.Flag{
 		utils.AccountSigSchemeFlag,
 		utils.AccountSetDefaultFlag,
 		utils.AccountFileFlag,
+		utils.AccountLabelFlag,
 	}
 	addFlags = []cli.Flag{
 		utils.AccountQuantityFlag,
@@ -67,6 +69,7 @@ var (
 		utils.AccountPassFlag,
 		utils.AccountDefaultFlag,
 		utils.AccountFileFlag,
+		utils.AccountLabelFlag,
 	}
 	fileFlags = []cli.Flag{
 		utils.AccountFileFlag,
